Adds stakeholder, cuf and nsu packages
Added some common types used when dealing with sefaz. All three packages were
created inside the sefaz package.

Package stakeholder

Stakeholder is a type that represents a stakeholder in a DFe or an sefaz
interaction. It could be a CPF (Pessoa Física) or a CNPJ (Pessoa Jurídica). The
package contains validators and json marshalers

Package NSU

NSU is a number used by sefaz's dfe distribution service. It's an offset to
request new DFes from sefaz and we store this value alongside with the
documents.

Package cUF

The cUF is the official  brazilian state codes. This package provides validation
and serialziation.