Improves trace package to use Span
This commit defines a new Trace model to be used. This model has a
TraceID and a ProbabilitySample, so, it is possible to create and use a
Span and make a Trace View.

READEME was updated to explain how to use the new trace package.

Some old methods in the package was kept and will be deprecated soon.