Use "ensureTraceNotEmpty" in "WithTrace"
If the context has a empty trace, a new one will be created
only if a span started. So, if the system do not use a span, a empty trace will remain in context.

This commit use ensureTraceNotEmpty in WithTrace method. Now, every time
a trace is put in context, it will be not empty.