Adds Retrier Package
The retrier package is a reusable package to aid with adding retry
behavior to specific functions. It can be configured by means of a
Settings structure passed to New function, which will replace zero
values with sane defaults.

The retries uses a Backoff Calculator, a Retry Evaluator and an
Error Wrapper as strategies to its behavior. One basic implementation
of each is available, with their own constructors, settings structures
and sane defaults.

New commonly used strategies are highly encouraged to be pushed to this
package as well.