Adds embeded admin HTTP server into app
All apps now must export the /metrics, /healthy and /ready endpoints
over port 9000. This is intended as the new default for arquivei's
kubernetes configuration.

This change automatically starts an HTTP server when the app is created.

The /healthy endpoint starts as OK and returns error only after the
shutdown is completed.

The /ready endpoint starts returning error. It returns ok only when the
main loop starts and the app should be properly working. During a
gracefull shutdown it returns error.

It was also added public functions to handle externally both states for
the default app. An application could create it's own checker that
updates the /ready state as it goes.

Also a change was made in the RunAndWait functions to not return any
error. This functions should the the last thing called on the main
function and it already handle errors by logging.