Changes app's shutdown to use priority queue
Previously the shutdown handlers were handled in a slice and were
executed in the order they were added.

Now it was changed to use a container/heap implementation. Each shutdown
handler receives an optional ShutdownPriority option and are processed
from the highest to lowest during a graceful shutdown.

There is no order guarantee for handlers of the same priority.