Add legacy app shutdown handler
To avoid problems with current code runing the app without shutdown
handler priority, if no priority is ever given, it set's a decreasing
priority each time the register functions is called. This recreates the
previous behavior of a fifo for the shutdown handlers.

If the developer mixes handlers with and without a priority the app
panics. This was done to avoid confusion during code migration to the
new handlers with priority.