Changes shutdown handler constructor and pop
Drops the New function for creating the handlers and forces
the developers to manually create the scruct and pass it to the register
function.

Makes the shutdown handler respect the insertion order for items with
the same priority.

This breaks compatibility with previous behavior but it's easy to
migrate because the pop method will try to use priority but if none if
set it will still retain the previous behavior of respecting insertion
order.