[App] Improves shutdown timeout error messages
The shutdown procedure was using ctx.Err() as an erro messages when
detecting timeouts and deadlines throught the context. Sometimes, these
errors being logged into a log system resulted in readers confusion,
alerting that a service Request was somehow causing a timeout that was
killing the whole service.

The error messages were changed to be more specific and less generic,
in order to reduce the chance of confusion and speed up analysis when
tracking down issues.

Also, fixed some linter errors and executed a spell check tool in the
app package.