Add a shutdown call when app finished by itself
When a app.mainLoop ends by itself, only the Ready status is set to
false, unlike the others cases, where the shutdown method is called. The
goduck workers ends a lot by itself and they need a propely shutdown.

Adds a shutdown call in case when the app ends by itself.