Adds debug handlers to app admin interface
This change adds the go's pprof handlers by defautl on the admin
interface. This is useful for debugging in productions and this
handlers should not cause performance issues.

I also added some good old dump handlers. We used in the past to dump
the content of go's runtime.

The new URLs are:

    /debug/pprof/
    /debug/pprof/cmdline
    /debug/pprof/profile
    /debug/pprof/symbol
    /debug/pprof/trace
    /debug/dump/goroutines
    /debug/dump/memory
    /debug/dump/memstats