Adds stale middleware
This is a propose for an Endpoint middleware that marks the app as
unhealthy if no new request arrives until the defined timeout.

A new request would mark the app as healthy.

The idea is allow applications to be restarted by kubernetes if they
stop receiving requests. It's useful if your endpoint is executed by a
worker and the worker is frozen.