Adds logging middlewares
Added an endpoint logging middleware to be used with go-kit. The
middleware is configurable provided with a default config struct.

There is also two http middlewares to be used with gorilla/mux that
are supposed to be used with the logging middleware. One adds tracking
information (RequestID and Trace information) and the other adds extra
HTTP information in the context for the logging middleware to log.