Adds goduck helper to create new processor
GoDuck is a project that helps with consuming kafka and pubsub queues.
It has it's own interface so we only need to code the functions that
process the messages.

This commit adds a helper functions that converts a go-kit endpoint to a
goduck processor. This is useful so we can reuse our own endpoint
middlewares with go-duck.

Links:

 - https://github.com/arquivei/goduck/