Adds custom enrich functions for loggingmiddleware
Adds two new configuration to the logging middleware:

 - EnrichLogWithRequest
 - EnrichLogWithResponse

They are functions that are executed before and after the request. They
receive a zerolog.Context and should return the same Context with data
form request or response.

For convenience the EnrichLogWithRequest also returns a context.Context
so it is possible also to enrich the context.Context with custom data.

Also the EnrichLogWithResponse receives an error. Errors are already
handled by the middleware, but is passed here for special cases. note
that the default handling of errors will still occur.

The DefaultConfig struct was changed to a NewDefaultConfig function to
protect the default configs against editing.