Adds go-kit metrics middleware
This commits introduces the Metrifier and a go-kit metrics middleware
that uses the Metrifier.

The Metrifier is a struct that helps calculating prometheus metrics for
an arbitraty piece of code.

The go-kit metrics endpoint uses de Metrifier to calculate latency and
count for endpoints. It allows for adding extra matrics labels by using
decoders.

The logging middleware was changed to be initialized the same way as the
metrics middleware for consistency.