Fix metrifier and mertrics middleware
Last commit was incomplete and labels were not beeing applied correctly
by the middleware.

A bug in which the tags were not being passed from the middleware to the
metrifier was fixed.

The endpoint is now properly initialized and checks the label decoders.

The code was refactored to be split in more functions and documentation
was enhanced.