New DontPanic middleware and better panic handling
New DontPanic middleware recovers from panic and returns a new
errors.Error properly initialized.

A new function were added in the errors package to handle panics:
DontPanic, and a new Error constructor to be used with the
result of the recover() function.

This commit also adds better panic handling in the other middlewares.
They will recover, handle the panic as an error, and raise the same
panic again. This is useful so we not loose logging or metrics
information during panics.