[httpcomm] Add CommunicateWithJSONDetailed function
The CommunicateWithJSONDetailed function has been added to the
httpcomm package. This new function returns details of the HTTP
Response, such as Status Code and HTTP Headers.

Since HTTP Headers implementation is unlikely to change, the Headers
from the official http package is returned. This reduces the need of
writing new code, but has some risk of breaking in the future.

The CommunicateWithJSON was not previously returning Request ID. The
Resquest ID has been added to the CommunicateWithJSONDetailed implementation.
The CommunicateWithJSON signature was, for now, left unchanged.