Refactor app and better probes (#75)
* Refactor app and better probes

Problem: The way probes were implemented was too simple, being only an
boolean variable for each probe. We couldn't manage the state of the app
from multiple places, like have a middleware to control readiness by
error count and another middleware that controls it by average latency.

Solution: The probe was changed to be a group of probes and it's final
state is determined by the state of each probe. The probes are
independent and if the group is not ok, all failed probes are returned
in the cause return parameter.

The stale middleware was altered to make use of this new feature.

The app and default app had to be refactored to allow for creating an
app before applying the middlewares. This also allows for a better
control flow in the startup process. Now the app can be created early
and shutdown handlers and probes may be registered before calling the
main loop.

Unfortunately this is a breaking change and applications will have to be
refactored to be able to take advantages oh this new and better
control-flow.

Links:

 - #68

* Added documentation for app and probes

Added a godoc for the app package explaining the basics and also added
an example program using the app package.