Update dependencies
The command `go get -u` was run for the dependencies listed in the
`go.mod` file.

The `app/config.go` file needed to be refactored to due to changes made
to the uconfig.Files type.

Linters `interfacer` and `maligned` were removed because they were
deprecated in golangci-lint 1.38.0 (it's currently on 1.41.1).