Makes contextmap go-routine safe
Problem

If the contextmap is used by different go-routines sharing the same
context or parent context, with an existing contextmap, it will panic
due to concurrent writing on the map.

Solution

The contextMap type was changed to a struct with a RW mutex and the map.
All functions using the map now locks the contextMap with the
appropriate lock.

The commit also bumps all dependencies with `go get -u ./...`.