[trace] Creates trace/v2 using opentelemetry
Problem:

Trace's current library is using OpenCensus, which is not a standard among programming languages. Therefore, we must have a version of the Trace library that uses OpenTelemetry, which incorporated OpenCensus and is now the Trace standard.

Solution:

Creates a new (v2) version of the Trace library using OpenTelemetry. It also creates a usage example and a README explaining step by step how to use it.