Allows for more logging levels
We were parsing less logging levels that the zerolog package provides.
This is an older code and zerolog supports more levels now and has a
parsing functions.

Change how we parse levels to use zerolog's parsing function so we
support the same logging levels that they do.

Also adds trace level logging for the app package, so it's easier to
understand what the app package is doing through the logs.