Fix apiutil.ParseError (#97)
The ParseError function extracts an error code and description to be
used on a HTTP response. The problem is that the default function for
getting an HTTP status code considers SeverityInput as a Bad Request but
the ParseError sets the error code as ErrCodeInternal. This caused
confusion for users in some APIs.

The ParseError was altered to return ErrCodeBadRequest when there is no
code but the severity is SeverityInput. So both HTTP Status Code and
Error Code are consistent.