Metrics middleware v2
This is a proposal on how can we change and extend the metrics
middleware to handle more cases.

The Labels decoder was changed to an interface that also returns the
list of labels supported, so the middleware automatically configures the
internal metrifier.

The config now works as a builder so it can be used like this
`NewDefaultConfig().WithX().WithY()`.

There is a new option that is called ExternalMetrics. This is a
function that can be set and will be called passing the request,
response and error. This is done after the internal metrifier computes
the metrics.