[hotfix] Stringsutil Truncate
Problem:

The Truncate method of the "stringsutil" package did not consider
unicode characters whose size is greater than a conventional
character. For example, the character 맱. When transforming the
string into a slice of runes, that character causes the length of
this slice to be smaller than the length of the original string,
resulting in "slice bounds out of range".

Solution:

- Checks the size after converting to rune

- Adds a Fuzz test