Update splitio dependency to v6.1.5
The Splitio Client package that was being used as a dependency was stuck
at an old version. The imported commit was pre-go.mod, to a version of
the Splitio Repository that still used gopkg.toml.

With 'go mod', a module name was added with the 'v6' suffix, causing an
inconsistency with the imports on source code, and newer versions of the
Splitio package. The dependecy has been updated by both updating the
go.mod and updating the imports in the source code.