[loggingmiddleware] Fix poor Zerolog usage (#129)
Problem:

The loggingmiddleware in gokitmiddlewares uses the Update Context
capabilities from Zerolog loggers. However, some systems that updated
Zerolog to version 1.27.0 had log corruption.

Solution:

Upon investigating changes in Zerolog v1.27.0, it was found that Zerolog
changed the behavior of the function WithContext, changing it's receiver
from a pointer to a copy - and then making needed adjustements.

The loggingmiddleware uses a function 'initLogger' which would register
a logger into the received context and then return both the logger and
context. The logger returned and the logger registered in the context
were the same but, with the Zerolog changes, now they are different
loggers. Since some memory is shared in Zerolog implementation, using
the returned logger and the logger in the context results in data
overwritting and thus logs corruption.

The problem has been fixed by only returning the context with the added
logger, forcing the middleware to fetch the logger from the context.