Creates APP v2
This is the next iteration of package app. This package was reworked to
be more simple to use and easy standardization of apps.

Gracefulshutdown was refactored to take advantage of the new
signal.NotifyContext.

Now there is a commom configuration struct that can be embed by
applications and will help bootstrap the application in one step.

Trace is now configured by the bootstrap.

The admin server is now configurable and features can be turned off,
like the debug endpoints.

Graceful shutdown is now configurable.

RunAndWait now passes a context to the main loop func that is canceled
on Shutdown.

Recover from panic has been reworked. Now RunAndWait will catch a panic
and do a proper shutdown. Panics outside RunAndWait can be handled by
app.Recover().

More examples were added and some documentation was rewritten/fixed.