Adds wrap support to foundationkit error
Since go 1.13, go has a new way of handling wrapped errors, with the functions
errors.Is, errors.As and errors.Unwrap.

Currently, there is no easy way of extracting the root of an error. This could
be useful for example: If we wrap a library that returns an error with
enriched information, we could use `errors.As` to extract information about
what happened internaly.