Adds wrap support for errors.GetCode and GetSeverity
Since the last change that added support for `errors.Is` and `errors.As`
functions, it is possible to retrieve specific error types thar are wraped
with foundationkit error.

But in the opposite case, when we wrap a foundationkit error with a custom
error type (e.g. when adding custom information to the error that could be
retrieved), we loose the current error's Code and Severity information.

Using `errors.As` instead of a simple cast `.(Error)` solves that problem.

This commit doesn't change the `GetRootError` and `GetRootErrorWithKV`
functions since the semantics become tricky when dealing with mixed errors.