Fix typos and shutdown handlers kept hanging
Fixed some typos in comments.

During the execution of the shutdown handlers, when the context was
closed the functions emitted an error but continued processing. This
would make the next iteration to also emit an error but the app stops
at the first error and the shutdown functions would be left hanging.