Fix possible concurrency problem with probe groups
ProbeGroup uses a map to store probes. If a check was triggered
while a new probe was being added the code would panic. This problem
exists in v1 but it was never observed, propably because apps normally
start fast and kubernetes has a initial wait period until it starts
polling the probes.

A RWMutex was added co control access to the map.

Other parts of the code were refactored to fix typos and have more
semantic names.