Removes external dependencies from app/v2
We are planning on slitting app/v2 into its own repository. To do this
we want first to remove any dependency with foundationkit to avoid cyclic
dependency but also to avoid so many renovate bot requests because
foundationkit depends on too many packages.

The only the necessary parts of the log package were copied to
app/logger so the app can auto configure the log package. Package
fkit/errors were dropped in favor of a pure std package approach. The
trace package was also dropped and will no longer be included on the
app/v2.

Now app/v2 don't depends on any of foundationkit's packages and also
have a minimal external dependencies. Only zerolog, uconfig and prometheus.