Fixes and refactors trace v2
The trace v2 package was not working properly. It didn't support naming
the system and the integration with another packages were broken, e.g.,
not being injected on the logs.

I tried to solve all the issues without breaking too much compatibility
but some breaking changes on trace v2 were made.

The tracking middleware were coupling request and trace packages. I
split the middleware into two (the old one still exists but now is
deprecated), so request and trace has each one its own middleware. The
request middleware should be used after the trace middleware so it can
inject the request-id into the trace.

The previous middleware supports both versions of trace so it can be
used during the phasing out of v1.

The span object in trace v2 was dropped in favor of the otel span
directly. The otel span is already an interface and it is very
powerful. If we would wrap this interface we would need to write too
much code and still would lose some of its power. Because otel is a good
market standard I felt confident in exposing it's span and use the trace
v2 package only as an orquestrator.

The example was also redone and it's now complete dockerized and using
jaeger to collect and serve the traces.

A new gokit endpoint middleware was created that starts a trace for the
endpoint but also handle errors and allows for custom attributes to be
inserted on the span.