[httpmiddlewares] Create new HTTP Audit Middleware
Create a new http middleware called audit middleware.

The purpose of this middleware is to allow creating and pushing a customized message for every http request received.

It could be useful when we want to store requests information based on some parameters received or based on the status code returned.

The parameters availabe for creating a message is the http request itself and the response status code.

To create this middleware, user should provide an implementation of both `Exporter` and `CreateAuditMessageFunc`.

Example of usage:

```
auditMiddleware := auditmiddlewate.NewAuditMiddleware(
		getNotificationQueue(),
		myhttp.CreateEndpointWasRequestedMessage,
	)

	r.Use(
		gziphandler.GzipHandler,
		trackingmiddleware.New,
		enrichloggingmiddleware.New,
		auditMiddleware.New,
	)
```

Depending on how the CreateAuditMessageFunc is implemented, the Push function can return an error. For example the scenario when the request body is too large. Cases like this this should be treated in the implementation of `CreateAuditMessageFunc`