PR fixes
The Dockerfile was dropped and the docker-compose was enhanced due to
security issues raised by sonar.

Custom configuration was dropped in favor of the official environment
variables.

README as converted into a proper godoc.

Added some sanity checks for checking for missing environment
variables when the default OpenTelemetry values are not ideal.