Update uconfig to the retracted v1.2.1 version
The insane generation of retracted versions without code of uconfig
allied to the Renovate Bot insisting on updating it even if it is in
ignoreDeps have proven to be formidable adversaries.

We can't win agains them.

We join them.

All hail uconfig v1.2.1 the new lord. Now please, let us work in piece =P