Refactors errors package
The errors.E() function receives a slice of any arguments. This makes this
functions very easy to break or cause confusion because each argument
must be type casted and processed.

Refactors errors.E() to receive the error as the first argument and a
list of options. The Code, Op, Severity and KeyValue are now an
ErrorOption.