Allows trackingmiddleware to work with both Trace V1 and V2
Any format in which we receive the trace, whether v1 or v2,
we must forward it in both the new and the old format.

If we receive it in both formats, the v2 format will be
prioritized.

This enables the migration from trace v1 to trace v2
without breaking compatibility.